see desc
We can now simulate all of the ammo and armor in the game 🥳. On top of this the armor damage calculation has been improved. STK has been renamed to HTK as it's a better name, and I've set it to round rather than floor on a fractional number, DAM is now DMG, enabled advanced filtering on the AEC and a few other small interface changes.